Move fake_haproxy.go to its own package
